refactor: update stateVersion filename
